Altered each pom.xml to successfully build (with warnings) under Maven 3
The sourceforge version of robocode builds under Maven 2, which allows
profile.xml to set the value ${robocode.version}, which was then used
in the pom.xml files.  Maven 3 does not support this, and recommends
the values be moved into settings.xml.  So that students don't have to
set up settings.xml, instead I've just globally replaced the
robocode.version variable with the value 2012.0.1-SNAPSHOT, which we'll
use for the course.

I've also removed the plugin version numbers in super-pom/pom.xml,
which had been set to plugins for running under Maven 2. In the next
commit, I'll set those to values for Maven 3.