Added super-pom to top-level build order.
Previously, super-pom was not included in the top-level build order.
This means a mvn install on the robocode directory would not install
the super-pom pom.xml in your maven repository, and a subsequent
build of the (separate) robots project would fail because it would
be unable to resolve the super-pom when trying to resolve the
robocode.api artifact's dependencies. Adding the super-pom to the
build order ensures it will be installed in your Maven repository
when you next peform a top-level mvn install.