Changes to the Rule Class
Make rules dynamic rather than static, and add a number of butt methods
to the class for future implementation of robot specific rules.

This new dynamic passes a robot to the rules, and determines the rules
for that specific robot.  This then assigns the rules to a set of
variables.

This has obviously broken the code, we will look at fixing this in
another commit.  This commit is so we can work on it on our branch.

The other thing we may have to look at is making the rules only being
able to be retrieved by a get method, rather than public variables so
these cannot be changed.