Made RobotPeer Extendable
RobotPeer is now extendable. Team's may implement specific rule changes by extending the RobotPeer class and then overriding any methods which control the Robot's behaviour. For instance: to have all robots move at double the speed, you need to write a new peer class that extends RobotPeer, and overrides the getNewVelocity() method such that it returns double the default value.

In order to these changes to take effect though it is necessary for the Battle class to be changed so that it assigns the extended class in place of the RobotPeer class. That is not dealt with in this commit.