Added basic Mode functionality using the strategy pattern
Also created a demonstration of a "Slow Mode" which just divides the
robots' velocity by two in RobotPeer. The Battle.setup() simply creates
a new instance of ClassicMode (change this to SlowMode and restart to
see it in action). Some problems with this is it requires the Interface
to contain every method that every mode wants to implement.