Added empty methods for override of RobotPeer
Added a number of methods which override existing methods in RobotPeer
in the class StatRobotPeer.

These should be all the necessary methods to overwrite to ensure that
our robots with equipped items are able to use their maximum stats
granted, even if these are outside the normal rules bounds.

Each method states what stats it changes compared to the existing rules
and RobotPeer