Renamed StatRobot to StatsRobot, as per our plans
A StatRobot has multiple stats, and we planned to call it "StatsRobot" in
week 2 or 3 (see whiteboard image).

Also changed the class' comments to be more general - i.e., other teams may
want to use this robot (and all the associated interface/peer hierarchy) to
change stats in their own way.