Add attributes and equipment API
Make large changes to the equipment API and implementation developed in my
commits from earlier this evening:

* Rename and move EquipmentPart.Position to EquipmentPartSlot in a separate
    file.
* Made a RobotAttributes enum, and changed EquipmentPart (and its Builder) to
    use the values of this enum for keys to attribute modifiers, rather than
    strings. This is a much more resilient way of doing things; new attribute
    definitions no longer require accompanying getters and setters. It hence
    cuts a lot of code from EquipmentPart (and allows iteration through all
    attributes).
* Modify the Equipment class to use this new API.

Add an attributes HashMap to RobotPeer to associate values with every
attribute, thereby allowing a method for game mechanics to be modified on a
per-robot basis.

Add an equipment HashMap to RobotPeer to associate EquipmentPart objects to
EquipmentPartSlot values, so as to ensure that only one part can occupy a slot
at a time.

Add the equip method to AdvancedRobot, and down the call hierarchy to
RobotPeer, via AdvancedRobotProxy, and in the associated interfaces -
IAdvancedRobotPeer and IRobotPeer. The equip method in RobotPeer handles the
logic; it modifies the robots attributes HashMap according to the attributes of
the part. There's also an accompanying unequip method for RobotPeer to revert
the modifications provided by the part already in the slot.

Delete the StatsRobot and StatsRobotPeer classes, as these are redundant with
a working attributes implementation in RobotPeer (and hence all robots -- no
need for a new robot type).

Add a very basic test robot for testing the user-facing equipment API (i.e. the
equip method in AdvancedRobot down the call stack to RobotPeer).

Add a lot of comments all over the place, since this will soon be merged into
master, and I know other people will be using it soon after that.

Have fun!