Simplify and secure Equipment API
Changes the equip() method of AdvancedRobots to accept just a part name - a
string associated with a part, as defined in Equipment. By putting the
RobotPeer in charge of getting the equipment part object, this ensures that
robot programmers can't define their own equipment parts, and then equip them.

It also removes the redundancy of passing in an EquipmentSlot parameter, when
the EquipmentSlot is already defined for that EquipmentPart. The disadvantage
of removing this is that the equip call is less descriptive, but it just means
we have to be more descriptive in the names of the part (i.e. make it obvious
what slot the part would occupy).