Finished modifying and changing the methods for ruleChecking on RobotAttributes
This commit has finished modifying the RoboPeer, RobotStatistics,
BulletPeer and Rules, to make sure that the game checks the robots
scaling attribute factors, rather than the hardcoded rules directly.

This commit finished making modifications for the following attributes:
Energy
Energy_Regen
Armor
Bullet_Speed
Ram_Attack
Ram_Defense

One method in rules was depreciated as it was moved to RobotPeer as it
needs to be accessed on a dynamic level, rather than static.

This should be the final commit regarding the rule checking, however a
few more final tweeks may need to be made.