Finished modifying and changing the methods for RobotPeer rule checking
This commit (i hope) has finalised the changes for the robot's rule
checking capabilities, and check the individual robots attributes
(factors) rather than the hard coded rules.

This means it checks attribute factor * appropriate base rule, rather
than the hardcoded base rule

This commit also had to change some things in BulletPeer and
RobotStatistics for this.

A method was depreciated (getBulletSpeed) in Rules, as it now needs to
be accessed dynamically because of the way it now checks for rules.  It
has since been moved to RobotPeer.

There may be a few more minor tweeks in future commits.

The attributes which checking was added for in this commit:
Energy
Energy_Regen
Armor
Bullet_Speed
Ram_Attack