Commit to fix breaking build due to RuleChecking Commits
This commit fixes the breaking build on team-forkbomb, due to the
previous additions to check for rules based on attributes, and not
hard-coded rules.

This commit also depreciated another method in Rules, as again it no
longer has a static value.

The changes to BulletPeer are reverting the previous commit, and have
to TODO items in there, that need investigations these were two of the
lines causing the broken build.

The changes to RobotPeer have added a number of methods to retrieve the
attributes of a robot (that is the factor of the attribute * the rule),
so a user can call these directly.  These methods are also used
throughout RobotPeer.  RobotPeer also has two TODO items to be
completed, as these were the other two items causing a broken build.

The last changes were made to ExecCommands which are basically now just
ensuring a robot can be set to a higher max velocity and turn rate,
which were missed in the previous commits.

In future commits the four TODO items will be corrected, as well as
further modifications to BaseRobotProxy, and SoundManager to enhance
rule checking and robot attributes.

It has also been noticed in this build (it is successful) but the test
"TestFairPlay" is based on statistics and as a result bit unreliable
(hence why it has broken master on jenkins previously as well as
"TestDuplicateScores".  I will be mentioning this on the newsgroup and
drop an email to Will.