Killstreaks now tracked and working. Not enabled by default.
At a killstreak of 5, a robot will now call an airstrike. To enable
killstreaks, set the enableKillstreaks flag in the new
KillstreakTracker class.

Killstreaks are not yet incremented for ram kills.

This method of tracking killstreaks is more of a prototype than a full
implementation, but is is simple, and avoids modifying RobotPeer for
the time being.