Final Changes to Rule Checking
This commit is fixing the four items that still had to be modified in
RobotPeer and BulletPeer that where checking the rules so robots with
attributes weren't accused of cheating.

This commit also removed the depreciation of the getBulletSpeed method
in the Rules, which was added in a previous commit.  To avoid
multiplying doubles this is used for robots with attributes of 1.0

There will be one more commit to change a few other thins regarding
this, such as sound, etc.  This will not affect the current rule
checking, and this should be complete and working.