Commit to fix the broken build caused: Still errors
This commit will have the build in a working state again.

The rest is up to the breaking team to fix.  The commit that broke the
build being b970a17a8636fa47f6f963e9fb6214f80b818951

It has numerous compilation errors, this was the one that was failing
the mvn install.

For more information see the chat of the breaking commit.