Create files for communication protocol
Create the required files for issue #28 (To be updated soon) which is a
communication protocol between robots. A *dispatch* is the term used to
refer to the objects/messages sent between robots.

`CommReceiveEvent.java` is an event that handles what happens when a
robot receives a dispatch.
`CommSendEven.java` is an event that handles what happens when a robot
sends a dispatch.
`DrawDispatch.java` draws the dispatches as 'speech bubbles'.
`SoldierRobot.java` a robot type that can receive command dispatches.
`ICommDispatch.java` an interface that sets the types and functions of
a dispatch.
`ICommReceiver.java` an interface that can be used by robots who wish
to receive dispatches.
`ICommSender.java` an interface that can be used by robots who wish to
send dispatches.