Added a HelicopterPeer using composition
This could work as a possible method of creating a custom robot to call
as the 7th killstreak. No functionality yet, just a skeleton for the
HelicopterPeer class. Still need to trace back up to the battle to
create a HelicopterPeer in the actual battle. This might be a problem
because a lot of references to RobotPeer are hard coded.