RuleChecking Final changes for robot Attributes
These were the last few things that still had references to the hard
coded rules and shouldn't have.  That is, some of the old robots, robot
proxys, etc. do, but these have no effect on the actual rules.

It should also be noted that the sound that uses the max_velocity has
not been changed, as this sound is the same for all robots (even if
they shoot a min powered bullet), thus it was left that way.