Removed reflection solution to try new solution, Items working with mode - need to test,
Reflection solution has been removed to test and add a new solution on
tying items to modes.

ClassicMode now implements two methods getItems and setItems.
setItems() is called in Battle in the initilising of the items. It goes
into the Mode and creates the items that it requires and will then set
the Battle's items to be the Modes items, whatever this may be.

This also means that if an item will be permanently on the board, it
can just be added to 'items' in Battle along with the Mode specific
items.