Disable debug print statements
In order to fix some of the failing tests as per ticket #42, I am
disabling print statements added in for debugging purposes. Some of
these can break the build when console output is being checked in a
test. The debug print statements can be enabled while testing code, but
should be disabled in the committed version.