Implemented Scoring
Score is incremented on each occurence of the ball bot entering the
goal area, the score is displayed at the end of 10 rounds (one point
for each goal). There is rank, team name and score, there are no other
categories presently. Need to get bounds for goal image to make more
accurate.