Basic teleportation appears to be working
Basic teleportation appears to be working. Bullet teleportation is not
working as of yet. No images are drawn yet. Please note that there are a
few issues:

1. In RobotPeer.java, line 1024. We too in our code removed the
@Override annotation but I am not sure what the outcome to removing this
in the master copy
2. This will probably be a bigger issue. I needed to create a method
stub for PerformMove with its original arguments to fullfil the
implementation of IRobotPeer and IRobotBattlePeer. I could have solved
this by creating PerformMove to be abstract. We need to dicuss this
issue as a group at the next meeting.