Added initial minion classes, integration. For issue #23.
Added the basic functionality for minions.
This includes integration into Battle - providing spawn functions that
can be called from any Robot. MinionProxy allows limited functionality
between a minion and it's parent.

As of this commit, UI integration and loading of minions has not been
completed - minions are simply copies of parents.

See relevant commit in robots for example spawning functionality.