Weapon Equipment Images not applied to all robots
getImage method now takes an imagePath string and now completly controls
whether a custom gun image is used. This control has been removed from
the getColoredGunRenderImage method. Previously when an equipment robot
was first in the robot queue, and applied an equipment image, this image
would be applied to all robots in the battle regardless of whther they
had that equipment item or were even equipment robots. This has now been
removed for the weapon images. Body and Radar fixes will be delivered
soon.