Body and Radar equipment images not applied to all robots
getBodyImage and getRadarImages methods now exclusively handle whether
or not custom images get applied to a robot. This removes the previous
issue where when an equipment robot applied a custom image when it was
first in the battle queue, the image was applied to all robots
regardless of whether they equipped that item.