Custom Highscore Support
Re-wrote just about everything in BattleResultsTableModel to do the
following:
1. Able to show different scores instead of them all
2. Able to choose a different order of scores
3. Able to set a different title
4. Ability to add new kind of scores to the highscore results easily

The new Scoring table is tied to a Mode and is able to be customised
through the new methods in ClassicMode that can be overrided.
Currently, ClassicMode has the default view of highscores.

FlagMode has been updated to add the scores that it requires.

SoccerMode has been updated to add the scores that it requires instead
of extending BattleResultsTableModel to do it itself.