Removed Resuable Code in BattleResultsTableModel and added Title Support
There was too much reusable code in BattleResultsTableModel for my own
liking (despite being my own code) so it's been removed.

There is now also support to change the title of a column for the
scoring. It is modified and can ONLY be modified on the call that sets
that score to appear.

FlagMode has had it's title Updated.