Fixed the BulletState property never being set to FIRED
When implementing my code to do with the sound functions of robocode I
noticed that the sounds associated with the FIRED state (ie, when the
bullet is fired) were never being played. I deliberated whether or not
to open an issue, decided to see how severe the problem was first. In
the end the problem was pretty simple, just took a heap of logging to
find. The code was changing the bullet state from FIRED to MOVING when
the frame was equal to 0, instead of 1.