New Test for EquipmentRobots also added more get methods for Advanced Robots
Added a new test for equipment robots to test the firepower and damage
taken of the robots, to ensure the attributes are working correctly.

Also added three methods robots can call to find the value of
attributes, being getMaxVelocity, getMaxBulletPower, getMinBulletPower,
as Rules can be redundant for some new robots.