Fixed an error with max volume
Previously the volume of a bullet was derived from the bullet power /
the max bullet power to get a value between zero and one. Now however
our items are causing the bullet power to exceed the maximum, which was
breaking things. Included an if statement to check if the bullet power
was exceeding the max bullet power and if so, return 1.