getFullEnergy method added to robotPeer, robotSnapshot
getFullEnergy method has been added to robotPeer and is now gathered as
a part of the robotSnapshot. fullEnergy is how much energy a robot can
possibly have i.e. what it has at the start of every round. By making it
available as part of the robotSnapShot class it can be used in
battleView.