Fixed Results Table Bug
The results table would appear after running a single button but after running
it again, it would not show and would throw a IndexOutOfBoundsException.

The Cause:
The Mode specific method to set the scores that were to appear in the table
were being added twice, also causing the column count to double in size. When
printing the results table, it would use the column count but it could not
reach the new scores added causing the IndexOutOfBoundsException.

The Fix:
The scores columns ArrayList is cleared before setting new scores to appear.