Overall Score Modifications, Bug Fixes and Table Upgrade to Flag Mode
1. No longer returns a HashMap to iterate over, returns a double of the
scores affected overall score.
2. No longer need a boolean on the show methods since the methods will
only be called if the score will be affected.
3. Expanding the names of the methods from abbreviations (keeping the
show methods despite there already being methods that do the same
thing, the naming of the functions make it easy to understand - might
modify later)
4. Fixed bug in BattleResultsTableModel getting a null instance and
calling clearColumns() on it for the first run of the Battle.
5. Flag Mode now only has Flag Score affecting the Overall Score so the
Overall Score is now displayed under the title 'Flag Score'.