Merge Fixes + Error Fixes (Waypoint Group)
In my last commit, forgot about RobotPoeer and IMode so these have both been
fixed for merging errors keeping both implementations from HEAD and master.

There were A LOT of errors caused by which ever group implemented the Waypoints
by not implementing the methods in the classes that implement the Events
Interfaces they modified.

They fix it for now but they all need to implement it.