Updated BattleResultsTableModel
Most of BattleResultsTableModel's methods now return itself so that a chain of
methods can be called on it in a chain-like format.

FlagMode, ClassicMode, SoccerMode and ZombieMode have all been updated to use
the chaining methods instead of individual calls.

There is no method now to show or remove the highscore from the table, once it
is added, there should be no time it will be deleted. Therefore, this option
has been removed. It may come in handy at a later date when it is the last
round.