Alter dynamic equipment implementation strategy
Rather than implementing an entirely new specification framework around
Equipment parts so they could be defined and selected for the battle in the
same way robots can, this changes the strategy to simply selecting an equipment
definition file, and then parsing the contents of that file to create an
EquipmentSet.

This commit does not implement that, but it does change the work done in commit
00820b8a60eacb41461aa0160ced853c0e5213f4 to use different equipment calls down
to the BattleSpecification.