Implement Equipment tab on new Battle dialog
Provides basic functionality in that the file selected in the Equipment tab
will propagate down to the EquipmentSet's named constructor "fromFile". Also,
selects a default definition file if none was selected, to allow for the tests
to continue functioning in future (when the definition file actually does
something).

Now, all that's left to do is implement the file parsing to create the parts
dynamically.