Modified test slighlty
I checked and put everything back, that i had changed in commit
5d944ed, to no avail, the output had changed by a value of 1 in score.
For such a small value, I presume there is a double someone that has
been rounded, where it no longer was, and thus changed the values by 1,
as it is not very different from original values.

This may not be optimal.  But because it is based on output, and scores
are originally sent in doubles, then rounded to an int, this may not be
found.